+++ pkg/domain/s3.go
@@ -1,4 +1,4 @@
-package core
+package domain
 
 import (
 	"context"
