+++ pkg/dynamodb/db_test.go
@@ -1,4 +1,4 @@
-package dynamoStore
+package dynamodb
 
 import (
 	"context"
@@ -7,7 +7,7 @@ import (
 	"github.com/aws/aws-sdk-go-v2/credentials"
 	"github.com/aws/aws-sdk-go-v2/service/dynamodb"
 	"github.com/aws/aws-sdk-go-v2/service/dynamodb/types"
-	"github.com/pennsieve/pennsieve-go-core/pkg/dynamoStore/models"
+	"github.com/pennsieve/pennsieve-go-core/pkg/dynamodb/models"
 	log "github.com/sirupsen/logrus"
 	"os"
 	"testing"
@@ -224,7 +224,7 @@ func TestMain(m *testing.M) {
 	os.Exit(code)
 }
 
-// populateManifestTable populates the test dynamoStore table with entries for testing
+// populateManifestTable populates the test dynamodb table with entries for testing
 func populateManifestTable() {
 
 	ctx := context.Background()
