+++ pkg/dynamodb/manifestFiles.go
@@ -1,4 +1,4 @@
-package dynamoStore
+package dynamodb
 
 import (
 	"context"
@@ -9,7 +9,7 @@ import (
 	"github.com/aws/aws-sdk-go-v2/feature/dynamodb/attributevalue"
 	"github.com/aws/aws-sdk-go-v2/service/dynamodb"
 	"github.com/aws/aws-sdk-go-v2/service/dynamodb/types"
-	"github.com/pennsieve/pennsieve-go-core/pkg/dynamoStore/models"
+	"github.com/pennsieve/pennsieve-go-core/pkg/dynamodb/models"
 	"github.com/pennsieve/pennsieve-go-core/pkg/models/manifest"
 	"github.com/pennsieve/pennsieve-go-core/pkg/models/manifest/manifestFile"
 	log "github.com/sirupsen/logrus"
@@ -20,7 +20,7 @@ import (
 
 // SyncFiles adds or updates files in the manifest file table
 func (q *Queries) SyncFiles(ctx context.Context, tableName string, manifestId string, fileSlice []manifestFile.FileDTO, forceStatus *manifestFile.Status) (*manifest.AddFilesStats, error) {
-	// Create Batch Put request for the fileslice and update dynamoStore with one call
+	// Create Batch Put request for the fileslice and update dynamodb with one call
 	var writeRequests []types.WriteRequest
 
 	var syncResponses []manifestFile.FileStatusDTO
@@ -29,7 +29,7 @@ func (q *Queries) SyncFiles(ctx context.Context, tableName string, manifestId st
 	var nrFilesUpdated int
 	var nrFilesRemoved int
 	for _, file := range fileSlice {
-		// Get existing status for file in dynamoStore, Unknown if does not exist
+		// Get existing status for file in dynamodb, Unknown if does not exist
 		var request *types.WriteRequest
 		var setStatus manifestFile.Status
 		if forceStatus == nil {
@@ -94,7 +94,7 @@ func (q *Queries) SyncFiles(ctx context.Context, tableName string, manifestId st
 			setStatus = *forceStatus
 		}
 
-		// If action requires dynamoStore actionm add request to array of requests
+		// If action requires dynamodb actionm add request to array of requests
 		if request != nil {
 			writeRequests = append(writeRequests, *request)
 		}
@@ -196,7 +196,7 @@ func (q *Queries) SyncFiles(ctx context.Context, tableName string, manifestId st
 	return &response, err
 }
 
-// UpdateFileTableStatus updates the status of the file in the file-table dynamoStore
+// UpdateFileTableStatus updates the status of the file in the file-table dynamodb
 func (q *Queries) UpdateFileTableStatus(ctx context.Context, tableName string, manifestId string, uploadId string, status manifestFile.Status, msg string) error {
 
 	_, err := q.db.UpdateItem(ctx, &dynamodb.UpdateItemInput{
@@ -309,7 +309,7 @@ func (q *Queries) GetFilesPaginated(ctx context.Context, tableName string, manif
 			}
 		}
 	case false:
-		// Query from main dynamoStore
+		// Query from main dynamodb
 		queryInput = dynamodb.QueryInput{
 			TableName:              aws.String(tableName),
 			ExclusiveStartKey:      startKey,
@@ -363,10 +363,10 @@ func GetWriteRequest(manifestId string, file manifestFile.FileDTO, curStatus man
 
 	// Switch based on provided status from client
 	// file --> provided as part of request
-	// curStatus --> current status in dynamoStore
+	// curStatus --> current status in dynamodb
 	switch file.Status {
 	case manifestFile.Removed:
-		// File is removed after being synced --> remove from dynamoStore if not uploaded already.
+		// File is removed after being synced --> remove from dynamodb if not uploaded already.
 		// If uploaded --> return current status
 
 		switch curStatus {
@@ -400,7 +400,7 @@ func GetWriteRequest(manifestId string, file manifestFile.FileDTO, curStatus man
 
 			return nil, curStatus, nil
 		default:
-			// If server synced or failed --> remove from dynamoStore
+			// If server synced or failed --> remove from dynamodb
 			data, err := attributevalue.MarshalMap(models.ManifestFilePrimaryKey{
 				ManifestId: manifestId,
 				UploadId:   file.UploadID,
@@ -450,7 +450,7 @@ func GetWriteRequest(manifestId string, file manifestFile.FileDTO, curStatus man
 
 			return &request, manifestFile.Verified, nil
 		case manifestFile.Registered, manifestFile.Failed, manifestFile.Unknown:
-			// server is synced, failed, unknown --> add/update the entry in dynamoStore
+			// server is synced, failed, unknown --> add/update the entry in dynamodb
 			item.Status = manifestFile.Registered.String()
 			item.InProgress = "x"
 
@@ -584,7 +584,7 @@ func (q *Queries) statusForFileItem(ctx context.Context, tableName string, manif
 
 	result, err := q.db.GetItem(context.Background(), getItemInput)
 	if err != nil {
-		log.Error("Error getting item from dynamoStore")
+		log.Error("Error getting item from dynamodb")
 	}
 
 	var pItem models.ManifestFileTable
@@ -601,7 +601,7 @@ func (q *Queries) statusForFileItem(ctx context.Context, tableName string, manif
 	return manifestFile.Unknown, nil
 }
 
-// RemoveFailedFilesFromResponse removes any files from the syncResponse that has not been successfully created in dynamoStore
+// RemoveFailedFilesFromResponse removes any files from the syncResponse that has not been successfully created in dynamodb
 func removeFailedFilesFromResponse(failedRequests []string, syncResponses []manifestFile.FileStatusDTO) []manifestFile.FileStatusDTO {
 
 	var newResponses []manifestFile.FileStatusDTO
