+++ pkg/dynamodb/manifestFiles_test.go
@@ -1,4 +1,4 @@
-package dynamoStore
+package dynamodb
 
 import (
 	"context"
