+++ pkg/dynamodb/models/manifestFileTable.go
