+++ pkg/dynamodb/models/manifestTable.go
