+++ pkg/dynamodb/store.go
@@ -1,4 +1,4 @@
-package dynamoStore
+package dynamodb
 
 import (
 	"github.com/aws/aws-sdk-go-v2/service/dynamodb"
